refactor(map): extract RF tool state and compute effects into hooks
