+++ frontend/src/pages/map/useUrlMapSync.ts
@@ -1,0 +1,100 @@
+import type { Map as MlMap } from "maplibre-gl";
+import { useEffect, useMemo, useRef } from "react";
+import { useSearchParams } from "react-router";
+
+import type { IMapNode } from "./types";
+
+/** Owns ?node= deep-link fly-to + debounced URL sync for center/zoom. */
+export function useUrlMapSync(
+  nodes: Record<string, IMapNode>,
+  mbMapRef: React.RefObject<MlMap | null>,
+) {
+  const [searchParams, setSearchParams] = useSearchParams();
+  const urlNodeId = searchParams.get("node") ?? "";
+  const urlNodeIdRef = useRef(urlNodeId);
+  urlNodeIdRef.current = urlNodeId;
+
+  const flyToTarget = useMemo(() => {
+    if (!urlNodeId) return null;
+    const node = nodes[urlNodeId] ?? nodes[`!${urlNodeId}`];
+    if (!node?.map_position) return null;
+    return node.map_position as [number, number]; // [lon, lat]
+  }, [urlNodeId, nodes]);
+  const flyToTargetRef = useRef(flyToTarget);
+  flyToTargetRef.current = flyToTarget;
+
+  const flyToHandledRef = useRef<string>("");
+
+  // Retry until map is ready
+  useEffect(() => {
+    if (!urlNodeId || !flyToTarget) return;
+    if (flyToHandledRef.current === urlNodeId) return;
+
+    const [lon, lat] = flyToTarget;
+
+    const tryFlyTo = () => {
+      if (flyToHandledRef.current === urlNodeId) return true;
+
+      const mbMap = mbMapRef.current;
+      if (mbMap) {
+        flyToHandledRef.current = urlNodeId;
+        mbMap.easeTo({ center: [lon, lat], zoom: 14, duration: 1200 });
+        setSearchParams((prev) => { prev.delete("node"); return prev; }, { replace: true });
+        return true;
+      }
+
+      return false;
+    };
+
+    if (tryFlyTo()) return;
+
+    const timers: ReturnType<typeof setTimeout>[] = [];
+    for (const delay of [200, 500, 1000, 2000, 3500]) {
+      timers.push(setTimeout(() => {
+        if (urlNodeIdRef.current !== urlNodeId) return;
+        tryFlyTo();
+      }, delay));
+    }
+
+    return () => timers.forEach(clearTimeout);
+  }, [urlNodeId, flyToTarget, setSearchParams, mbMapRef]);
+
+  // Debounced URL sync for center/zoom
+  useEffect(() => {
+    let timer: ReturnType<typeof setTimeout>;
+    const syncUrl = () => {
+      clearTimeout(timer);
+      timer = setTimeout(() => {
+        let lat: number | undefined;
+        let lng: number | undefined;
+        let z: number | undefined;
+
+        const mb = mbMapRef.current;
+        if (mb) {
+          const c = mb.getCenter();
+          lat = +c.lat.toFixed(5);
+          lng = +c.lng.toFixed(5);
+          z = +mb.getZoom().toFixed(2);
+        }
+
+        if (lat != null && lng != null && z != null) {
+          setSearchParams((prev) => {
+            prev.set("lat", String(lat));
+            prev.set("lng", String(lng));
+            prev.set("z", String(z));
+            return prev;
+          }, { replace: true });
+        }
+      }, 800);
+    };
+
+    const mb = mbMapRef.current;
+    if (mb) {
+      mb.on("moveend", syncUrl);
+      return () => { clearTimeout(timer); mb.off("moveend", syncUrl); };
+    }
+    return () => clearTimeout(timer);
+  }, [setSearchParams, mbMapRef]);
+
+  return { searchParams, flyToTargetRef };
+}
